+++ src/utils/index.ts
@@ -1,2 +1,3 @@
 export {serialize} from "./serialize"
 export {findHeadings} from "./findHeadings"
+export * from './nodeGuards'