+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - doc: add serialize() documentation
 - feat!: change findHeadings() api & refine implementation & doc it
 - feat!: improve getHeadingText()
+- feat: export nodeGuards
 
 ## v2.0.0
 - feat!: custom lexical serializer instead of using @lexical/headless, added README, added CHANGELOG, renamed files to camelcase